+++ CHANGELOG.md
@@ -23,7 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fix missing JS file for PatternFly Java J2CL
 
 ### Removed
- W
+
 - Mockito dependency
 
 ## [0.2.4] - 2024-04-26