+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added missing password type to the type mapping for `TextInput` (thanks @kolka213)
 
+### Upgrades
+
+- Bump Elemento to 1.6.2
+
 ## [0.2.5] - 2024-05-21
 
 ### Added